update modules, add updateTodos
